Setup URL forwarding with plain HTML files (#46)
We were using subdomain forwarding (calendar.fatiando.org) for some hard
to remember URLs. This unfortunately doesn't work with HTTPS, leading to
some broken links and ugly warnings from browsers about potential
danger. Replace those with a simple HTML redirection structure in
subfolders. For example, `fatiando.org/calendar` now redirects to the
Google calendar link. Same for `fatiando.org/slack` and
`fatiando.org/good-first-issue`. There are instructions in the README
for adding new forwards.
A caveat is that we now have to be careful with the subfolders we create
here. But that's probably not going to be a problem since this website
is intentionally not very elaborate (it links to other resources).