Add Ensaio and rework the front page (#55)
Added a section for Ensaio to the front page. Had to do some reworking to make
it fit nicely with the design. Ended up removing that little navigation section which 
was a bit awkward with the logos in different sizes. Also removed the 
EBGaramond font for the title since it was a bit of a waste and looked off when 
compared with the rest of the site.