Make a sphinx extension for fetching authors list (#58)
Made a sphinx extension (in `sphinx_extensions/authors/__init__.py`) for
fetching the authors for each project and making the cards. This
reuses and replaces the code in `conf.py`, with the advantage of the
download only happening when the page with the list is updated (so
building the docs is much faster now). The extension creates the
`fatiando-authors` directive which creates the list given the name of a
package. This is very specific to our projects so there's no point in
making it a separate package. The code is based on the
sphinx-contributors and takes some code from sphinx-panels extensions
(both MIT licensed). Our extension itself is also licensed MIT for
convenience.

Co-authored-by: Santiago Soler <santiago.r.soler@gmail.com>