Add a alert about our t-shirts
Includes the option to set a dismissable alert to the top of every page
from the `conf.py` file (using the `html_context` variable). Use it to
set an alert out the new t-shirts available from the Software
Underground store. Include this in the "Support" page as well and link
to that page in the very top of the index instead of "Cite" (since it's
more general and links to Cite).