Make an extension to add the latest versions
In the list of info for each package, we now add a link to the latest
version (including the version number). Before we had these as
shields.io badges at the bottom but they were a bit hidden and looked
out of place in the info list. The extension queries the PyPI API to get
the latest version and inserts a link to it with the version number as
the text.