Add our Mastodon account to the menu and Contact (#78)
Had to edit the template to include the `rel="me"` attribute in the
social media links. This is used to verify that the owner of the website
and social media account are the same person. Make the social media
section of the Contact page more generic ("Social media" instead of
"Updates" and use a different icon than the twitter bird). Move it up a
bit since it was kind of hidden as the last element of that page.